Cumulative commit: - Fixed skins not properly selected in player menu after selecting them when joining in team-based games like INF. - Fixed all skins not properly shown in player menu in H&S and H&Z. - Fixed skins not properly selected in player menu in H&S and H&Z. - Fixed gametype info screen not coming up after using the 1fx. HTTP Downloader and joining back in-game.
NOTE:
The approach for determing to use all skins in H&S/H&Z, even though it is a team-based game, is somewhat weird at first sight.
The normal approach would be to use something like a new ui_* CVAR to determine if all skins should be used or not,
manually editing the setup_player.menu afterwards. Instead everything is handled in-code and the gametype is manually checked.
This is done so all existing mod files do not need to be manually modified, and future mods could (in theory) be compatible
with this system without breaking other existing things.

As an extra bonus, this approach seems to properly update the selected skin in team-based games, using team or all skins.